Made schedule separate component
