+++ DoNotDisturb/ClientApp/src/app/inside/dashboard/dashboard.component.less
@@ -12,18 +12,6 @@
   position: fixed;
   bottom: 0;
 }
-.modal-fullscreen{
-  .modal-dialog{
-    max-width: 100vw !important;
-    height: 100%;
-    margin: 0;
-  }
-  .modal-content{
-    width: 95vw;
-    height: 95vh;
-    margin: 0 auto;
-  }
-}
 .alert-fullscreen{
   position: fixed;
   top: 0;
