+++ DoNotDisturb/ClientApp/src/app/outside/dashboard/dashboard.component.less
@@ -12,15 +12,3 @@ font-size: 2.5rem;
   position: fixed;
   bottom: 0;
 }
-.modal-fullscreen{
-  .modal-dialog{
-    max-width: 100vw !important;
-    height: 100%;
-    margin: 0;
-  }
-  .modal-content{
-    width: 95vw;
-    height: 95vh;
-    margin: 0 auto;
-  }
-}
