+++ DoNotDisturb/ClientApp/src/app/shared/components/schedule/schedule.component.html
@@ -1,0 +1,57 @@
+﻿<div #scheduleModal class="modal fade modal-fullscreen text-dark" id="modal-schedule" tabindex="-1" role="dialog" aria-hidden="true">
+  <div class="modal-dialog modal-dialog-centered" role="document">
+    <div class="modal-content">
+      <div class="modal-header border-0">
+        <div class="container-fluid">
+          <div class="row">
+            <div class="col-4">
+                <img (click)="goToSettings()" class="clickable" src="/assets/images/settings.png"/>
+            </div>
+            <div class="col-4 text-center">
+              <span class="h4">Upcoming meetings</span>
+            </div>
+            <div class="col-4 text-right">
+              <button type="button" class="close" (click)="hide()">
+                <span aria-hidden="true">&times;</span>
+              </button>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div class="modal-body">
+        <app-schedule-meetings *ngIf="!howToScheduleDisplayed" [meetings]="meetings"></app-schedule-meetings>
+        <div class="container" *ngIf="howToScheduleDisplayed">
+          <div class="row text-center">
+            <div class="col">
+              <span class="h2">How to schedule</span>
+            </div>
+          </div>
+          <div class="row">
+            <div class="col">
+              <div>Here are instructions on how to do it:
+                <ol>
+                  <li>Please visit your Google Calendar associated with your Shaw account</li>
+                  <li>Select the date and time for which you want to schedule your meeting</li>
+                  <li>Choose the desired room for the meeting to take place</li>
+                  <li>Fill out all necessary information required for the meeting</li>
+                  <li>Click Save, You have now scheduled your meeting!</li>
+                </ol>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+      <div class="modal-footer border-0">
+        <div class="container-fluid">
+          <div class="row text-center">
+            <div class="col">
+              <span style="cursor: pointer;" (click)="howToScheduleDisplayed = !howToScheduleDisplayed">
+                {{howToScheduleDisplayed ? 'Back to schedule' : 'Need to schedule a meeting?'}}
+              </span>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
