+++ DoNotDisturb/ClientApp/src/styles.less
@@ -16,3 +16,18 @@ body{
 .bg-black{
   background-color: #272727;
 }
+.modal-fullscreen{
+  .modal-dialog{
+    max-width: 100vw !important;
+    height: 100%;
+    margin: 0;
+  }
+  .modal-content{
+    width: 95vw;
+    height: 95vh;
+    margin: 0 auto;
+  }
+}
+.clickable{
+  cursor: pointer;
+}
